feat: add modpacks page with filtering and display
